removing attribute :name fixes 1.8.7 tests due to error: "comparison of String with :name failed occured"
Not sure where it originates from but the error is finicky.
When using a debugger, it "goes away". Most likely an issue
with ruby itself. The use of "attribute :name" in the template
is tested in many other tests so it's not needed here.
The two specs still test for the use of child.

Tests under 1.9.2 and 1.9.3 passed before and continue to.