Uses __send__ instead of send
rabl does not work well with decorators that use BasicObject as a
parent because BasicObject defines #__send__, not #send.

We ran into this with a gem called dumb_delegator that attempts to
delegate nearly everything to the underlying object, so you don't have
to do awkward things like explicitly delegate #class or #is_a? for
Rails. However, since Rabl uses #send, and not #__send__, messages are
not "overridden" correctly by the delegator.

A concrete example is in this gist:
https://gist.github.com/bb579e81d8353586d250