Added a simple LRU to Rabl::CacheEngine
CacheEngine `fetch` and `write` depend on `Rails.cache` and otherwise
drop caching on the floor. I don't know how much this affects
performance directly, as I was benchmarking the instance_eval at the
time and that was much coarser.

The LRU provided is perhaps too simple, but the idea is there.