Restyling / rephrasing some of the longer and more complex logic.
In most cases I simply reformat long lines to make the logic more
clear.

In some others, I was reshaping the logic in order to remove redundant
comments.

Others had simple styling issues or were using anachronisms.

This commit is entirely optional and I'm happy to pop it out or put it
in a different PR.